# RavenDB-22617 Corax's facet performance