--- a/src/Corax/Querying/Matches/Meta/IAggregationProvider.cs
+++ b/src/Corax/Querying/Matches/Meta/IAggregationProvider.cs
@@ -1,0 +1,11 @@
+using System;
+using System.Collections.Generic;
+
+namespace Corax.Querying.Matches.Meta;
+
+public interface IAggregationProvider
+{
+    public IDisposable AggregateByTerms(out List<string> terms, out Span<long> counts);
+    public long AggregateByRange();
+
+}