--- a/src/Corax/Querying/Matches/TermProviders/TermProvider.NumericRange.cs
+++ b/src/Corax/Querying/Matches/TermProviders/TermProvider.NumericRange.cs
@@ -10,6 +10,7 @@ using Sparrow;
 using Sparrow.Compression;
 using Sparrow.Extensions;
 using Sparrow.Server;
+using Sparrow.Threading;
 using Voron.Data.Containers;
 using Voron.Data.Lookups;
 using Voron.Data.PostingLists;
@@ -217,12 +218,11 @@ namespace Corax.Querying.Matches.TermProviders
             var allocator = _searcher.Allocator;
             
             NativeList<long> postingLists = new();
-            postingLists.Initialize(_searcher.Allocator);
-
+            postingLists.Initialize(allocator);
             NativeList<TermIdMask> postingListsType = new();
-            postingListsType.Initialize(_searcher.Allocator);
+            postingListsType.Initialize(allocator);
             
-            while (_iterator.MoveNext(out var termId))
+            while (_isEmpty == false && _iterator.MoveNext(out var termId))
             {
                 if (termId == _lastTermId)
                 {
@@ -248,9 +248,9 @@ namespace Corax.Querying.Matches.TermProviders
                 }
             }
             
-            using var _ = _searcher.Allocator.Allocate((sizeof(UnmanagedSpan*)) * postingLists.Count, out ByteString containers);
+            using var _ = allocator.Allocate((sizeof(UnmanagedSpan)) * postingLists.Count, out ByteString containers);
             var containersPtr = (UnmanagedSpan*)containers.Ptr;
-            
+      
             Container.GetAll(_searcher._transaction.LowLevelTransaction, postingLists.ToSpan(), containersPtr, singleMarker, _searcher._transaction.LowLevelTransaction.PageLocator);
 
             long totalCount = 0;