… src/Raven.Server/Documents/Queries/Facets/FacetedQueryParser.cs
… src/Raven.Server/Documents/Sharding/Operations/Queries/ShardedFacetedQueryOperation.cs
… src/Voron/Data/CompactTrees/CompactTree.Iterator.cs
… src/Voron/Data/CompactTrees/CompactTree.cs