### Issue link

https://issues.hibernatingrhinos.com/issue/RavenDB-22617

### Additional description

for aggregations like:
`facet(Field)` or simple numerical ones like `facet(Field >= Number and Field <= Number2)` we will use index to produce the output.

Results (10M docs, ~5M unique textual terms `(Tag)` , 100K unique numbers `(Number)`)
`select facet(Tag)`:
Corax (before): ~55000ms
Corax : ~8500ms
Lucene: ~12500ms

`facet(Number)`
Corax (before): ~19000ms
Corax: ~40ms
Lucene: ~1300ms

`facet(Number >= 100 and Number <= 10000)`
Corax (before): ~12000ms
Corax: <1ms
Lucene: ~200ms


### Type of change

- [ ] Bug fix
- [ ] Regression bug fix
- [x] Optimization
- [ ] New feature

### How risky is the change?

- [ ] Low 
- [x] Moderate 
- [ ] High
- [ ] Not relevant

### Backward compatibility

- [x] Non breaking change
- [ ] Ensured. Please explain how has it been implemented?
- [ ] Breaking change
- [ ] Not relevant

### Is it platform specific issue?

- [ ] Yes. Please list the affected platforms.
- [x] No

### Documentation update

- [ ] This change requires a documentation update. Please mark the issue on YouTrack using `Documentation Required` tag.
- [x] No documentation update is needed 

### Testing by Contributor

- [ ] Tests have been added that prove the fix is effective or that the feature works
- [ ] Internal classes added to the test class (e.g. entity or index definition classes) have the lowest possible access modifier (preferable `private`) 
- [x] It has been verified by manual testing

### Testing by RavenDB QA team

- [ ] This change requires a special QA testing due to possible performance or resources usage implications (CPU, memory, IO). Please mark the issue on YouTrack using `QA Required` tag.
- [x] No special testing by RavenDB QA team is needed

### Is there any existing behavior change of other features due to this change?

- [ ] Yes. Please list the affected features/subsystems and provide appropriate explanation
- [x] No

### UI work

- [ ] It requires further work in the Studio. Please mark the issue on YouTrack using `Studio Required` tag.
- [x] No UI work is needed