@maciejaszyk 2024-07-24
> What are the perfomance differences?

@ayende measurements are available in the description