@ayende 2024-07-24
@maciejaszyk where are we spending the time in the `facet(Tag)` scenario?