@ayende 2024-07-24
So the issue is simply the time that it takes us to decode because of the number of tags?

What if we have more reasonable number (like < 100 tags), what is the cost then in Croax 7 Lucene?