@maciejaszyk 2024-07-24
> So the issue is simply the time that it takes us to decode because of the number of tags?
> 
> What if we have more reasonable number (like < 100 tags), what is the cost then in Croax 7 Lucene?
query: `select facet(Tag)`
10M docs, 5150 tags:
Corax (before): ~13500ms
Corax (this PR): ~4ms
Lucene: ~1000ms

10M, 100Tags:
Corax (before): ~13200ms
Corax (this PR): <1ms
Lucene: ~650ms